Narrow ComponentID to UInt16 (draft-09)
Track draft-ietf-mls-extensions-09's uint16 ComponentID instead of -08's
uint32. The Exporter Tree's 2^16 leaves are exactly the 16 bits of a
ComponentID, so every id names a valid leaf — the out-of-range check and
ExportError.invalidComponentID are gone. swift-mls stays -09-clean; the
deployed fork's u32 wire is a downstream compat concern, not carried here.

+++ Sources/MLSKeySchedule/ComponentID.swift
@@ -1,35 +1,37 @@
 import MLSCodec
 
 extension MLS.KeySchedule {
-	/// draft-ietf-mls-extensions-08 §4.4's `ComponentID` — the identifier that
-	/// indexes the Exporter Tree (`SafeExportSecret(ComponentID)`).
+	/// draft-ietf-mls-extensions-09 §4.1's `ComponentID` — "a two-byte value that
+	/// uniquely identifies a component"; `uint16 ComponentID`. The Exporter Tree
+	/// (§4.4) indexes its 2^16 leaves by one.
 	///
-	/// draft-08 types it `uint32`, while giving the tree only 2^16 leaves;
-	/// draft-09 resolves that by narrowing `ComponentID` to `uint16`. We track
-	/// -08 and match the deployed fork (`type ComponentID = u32`, which keeps the
-	/// u32 type and rejects ids ≥ 2^16): so this is `UInt32`, and
-	/// `ExporterTree.safeExportSecret` rejects any id outside the tree's leaf
-	/// range rather than truncating it onto a colliding leaf.
+	/// draft-08 typed it `uint32` while giving the tree only 2^16 leaves; -09
+	/// narrows it to `uint16`, which we track: a `ComponentID` is exactly the 16
+	/// bits of the tree's leaf range, so every value names a valid leaf and no
+	/// runtime range check is needed. The deployed fork still encodes `u32` on the
+	/// wire (it tracks -08); reconciling that is a **downstream** concern — swift-
+	/// mls stays -09-clean and never carries the u32 form.
 	///
-	/// This is the draft-level *type* only. Specific component assignments are
-	/// the adopter's — e.g. an application's private-use ids — declared as
-	/// `static let`s on this type at the layer that owns them; none are
-	/// IETF-registered, so this type defines no cases.
+	/// This is the draft-level *type* only. §7.5 registers component types
+	/// (`app_components`, `safe_aad`, `content_media_types`,
+	/// `last_resort_key_package`, `app_ack`) but swift-mls implements none of them,
+	/// and an application's own ids are the adopter's — declared as `static let`s
+	/// at the layer that owns them — so this type defines no cases.
 	public struct ComponentID: Sendable, Equatable, Hashable, RawRepresentable {
-		public let rawValue: UInt32
+		public let rawValue: UInt16
 
-		public init(rawValue: UInt32) {
+		public init(rawValue: UInt16) {
 			self.rawValue = rawValue
 		}
 
-		public init(_ rawValue: UInt32) {
+		public init(_ rawValue: UInt16) {
 			self.rawValue = rawValue
 		}
 	}
 }
 
 extension MLS.KeySchedule.ComponentID: ExpressibleByIntegerLiteral {
-	public init(integerLiteral value: UInt32) {
+	public init(integerLiteral value: UInt16) {
 		self.rawValue = value
 	}
 }
